requirements.txt
albucore
albumentations
annotated-types
blinker
certifi
charset-normalizer
click
coloredlogs
contourpy
cycler
Cython
dlib
easydict
face-recognition
face_recognition_models
faiss-cpu
filelock
Flask
flatbuffers
fonttools
fsspec
huggingface-hub
humanfriendly
idna
ImageHash
imageio
insightface
itsdangerous
Jinja2
joblib
kiwisolver
lazy_loader
MarkupSafe
matplotlib
mpmath
networkx
numpy
onnx
onnxruntime
opencv-python
opencv-python-headless
packaging
pillow
prettytable
protobuf
pydantic
pydantic_core
pyparsing
python-dateutil
PyWavelets
PyYAML
regex
requests
safetensors
scikit-image
scikit-learn
scipy
simsimd
six
stringzilla
sympy
threadpoolctl
tifffile
tokenizers
torch
tqdm
transformers
typing-inspection
typing_extensions
urllib3
wcwidth
Werkzeug
